requirements.txt
aiohttp==3.9.3
fastapi==0.109.2
pydantic==2.6.1
python-dotenv==1.0.1
starlette==0.36.3